chore: remove .prettierrc
We failed to git add both the old and new names of the file before the last commit
so that git could track the rename. This commit removes the old version of the file,
but unfortunately, we will lose the history of the file. Not much history on it anyway
so I'm not going to bother to roll back the last commit and do it again.